Upgrade to version 0.1.1
