+++ lib/bashtetikn/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Bashtetikn
-  VERSION = "0.1.0"
+  VERSION = "0.1.1"
 end
